+++ package.json
@@ -2,23 +2,6 @@
   "name": "carbon",
   "devDependencies": {
     "@biomejs/biome": "2.3.8",
-    "@carbon/auth": "workspace:*",
-    "@carbon/config": "workspace:*",
-    "@carbon/database": "workspace:*",
-    "@carbon/documents": "workspace:*",
-    "@carbon/ee": "workspace:*",
-    "@carbon/form": "workspace:*",
-    "@carbon/jobs": "workspace:*",
-    "@carbon/kv": "workspace:*",
-    "@carbon/lib": "workspace:*",
-    "@carbon/locale": "workspace:*",
-    "@carbon/logger": "workspace:*",
-    "@carbon/notifications": "workspace:*",
-    "@carbon/react": "workspace:*",
-    "@carbon/remix": "workspace:*",
-    "@carbon/stripe": "workspace:*",
-    "@carbon/tiptap": "workspace:*",
-    "@carbon/utils": "workspace:*",
     "@lingui/cli": "catalog:",
     "@react-router/dev": "catalog:",
     "@types/aws-lambda": "8.10.152",
