+++ packages/auth/package.json
@@ -46,10 +46,15 @@
   "sideEffects": false,
   "version": "0.0.0",
   "dependencies": {
+    "@carbon/documents": "workspace:*",
+    "@carbon/env": "workspace:*",
+    "@carbon/kv": "workspace:*",
+    "@carbon/lib": "workspace:*",
+    "@carbon/logger": "workspace:*",
     "@react-email/components": "catalog:",
     "cookie": "catalog:",
     "dompurify": "catalog:",
     "nanoid": "catalog:",
-    "zustand": "4.5.7"
+    "zustand": "catalog:"
   }
 }
