+++ packages/auth/src/config/env.ts
@@ -1,432 +1,1 @@
-import { Edition, isBrowser, parseBoolean } from "@carbon/utils";
-
-declare global {
-  interface Window {
-    env: {
-      AUTH_PROVIDERS: string;
-      CARBON_EDITION: string;
-      CARBON_API_URL: string;
-      CLOUDFLARE_TURNSTILE_SITE_KEY: string;
-      CONTROLLED_ENVIRONMENT: string;
-      ERP_URL: string;
-      JIRA_CLIENT_ID: string;
-      MES_URL: string;
-      NOVU_APPLICATION_ID: string;
-      NOVU_API_URL: string;
-      ONSHAPE_CLIENT_ID: string;
-      POSTHOG_API_HOST: string;
-      POSTHOG_PROJECT_PUBLIC_KEY: string;
-      SUPABASE_URL: string;
-      SUPABASE_ANON_KEY: string;
-      VERCEL_URL: string;
-      VERCEL_ENV: string;
-      QUICKBOOKS_CLIENT_ID: string;
-      XERO_CLIENT_ID: string;
-      DEFAULT_LANGUAGE: string;
-    };
-  }
-}
-
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      CARBON_EDITION: string;
-      CARBON_API_URL: string;
-      CLOUDFLARE_TURNSTILE_SITE_KEY: string;
-      CLOUDFLARE_TURNSTILE_SECRET_KEY: string;
-      DOMAIN: string;
-      ERP_URL: string;
-      JIRA_CLIENT_ID: string;
-      JIRA_CLIENT_SECRET: string;
-      JIRA_OAUTH_REDIRECT_URL: string;
-      JIRA_STATE_SECRET: string;
-      MES_URL: string;
-      NOVU_APPLICATION_ID: string;
-      NOVU_API_URL: string;
-      NOVU_SECRET_KEY: string;
-      ONSHAPE_CLIENT_ID: string;
-      ONSHAPE_CLIENT_SECRET: string;
-      ONSHAPE_OAUTH_REDIRECT_URL: string;
-      POSTHOG_API_HOST: string;
-      POSTHOG_PROJECT_PUBLIC_KEY: string;
-      QUICKBOOKS_CLIENT_SECRET: string;
-      QUICKBOOKS_WEBHOOK_SECRET: string;
-      RESEND_API_KEY: string;
-      RESEND_DOMAIN: string;
-      SESSION_SECRET: string;
-      SESSION_KEY: string;
-      SESSION_ERROR_KEY: string;
-      SLACK_CLIENT_ID: string;
-      SLACK_CLIENT_SECRET: string;
-      SLACK_OAUTH_REDIRECT_URL: string;
-      SLACK_SIGNING_SECRET: string;
-      SLACK_STATE_SECRET: string;
-      STRIPE_SECRET_KEY: string;
-      STRIPE_WEBHOOK_SECRET: string;
-      STRIPE_BYPASS_COMPANY_IDS: string;
-      STRIPE_BYPASS_USER_IDS: string;
-      GTM_URL: string;
-      GTM_EVENTS_API_SECRET_KEY: string;
-      SUPABASE_ANON_KEY: string;
-      SUPABASE_URL: string;
-      SUPABASE_DB_URL: string;
-      SUPABASE_AUTH_EXTERNAL_AZURE_CLIENT_ID: string;
-      SUPABASE_AUTH_EXTERNAL_GOOGLE_CLIENT_ID: string;
-      SUPABASE_SERVICE_ROLE_KEY: string;
-      REDIS_URL: string;
-      VERCEL_URL: string;
-      VERCEL_ENV: string;
-      INNGEST_SIGNING_KEY: string;
-      INNGEST_EVENT_KEY: string;
-      XERO_CLIENT_SECRET: string;
-      XERO_WEBHOOK_SECRET: string;
-      DEFAULT_LANGUAGE: string;
-    }
-  }
-}
-
-type EnvOptions = {
-  isSecret?: boolean;
-  isRequired?: boolean;
-};
-
-export function getEnv(
-  name: string,
-  { isRequired, isSecret }: EnvOptions = { isSecret: true, isRequired: true }
-) {
-  if (isBrowser && isSecret) return "";
-
-  const source = (isBrowser ? window.env : process.env) ?? {};
-
-  const value = source[name as keyof typeof source];
-
-  if (!value && isRequired) {
-    throw new Error(`${name} is not set`);
-  }
-
-  return value;
-}
-
-/**
- * Server env
- */
-
-export const AUTH_PROVIDERS =
-  getEnv("AUTH_PROVIDERS", {
-    isRequired: false,
-    isSecret: false
-  }) ?? "email,google,azure";
-
-const CARBON_EDITION = getEnv("CARBON_EDITION", {
-  isRequired: false,
-  isSecret: false
-});
-
-const getEdition = () => {
-  if (CARBON_EDITION === "cloud") {
-    return Edition.Cloud;
-  }
-  if (CARBON_EDITION === "enterprise") {
-    return Edition.Enterprise;
-  }
-  if (CARBON_EDITION === "test") {
-    return Edition.Test;
-  }
-  return Edition.Community;
-};
-
-export const CarbonEdition = getEdition();
-
-export const CARBON_API_URL =
-  getEnv("CARBON_API_URL", {
-    isRequired: false,
-    isSecret: false
-  }) ?? getEnv("SUPABASE_URL", { isSecret: false });
-
-export const CLOUDFLARE_TURNSTILE_SITE_KEY = getEnv(
-  "CLOUDFLARE_TURNSTILE_SITE_KEY",
-  { isSecret: false, isRequired: false }
-);
-export const CLOUDFLARE_TURNSTILE_SECRET_KEY = getEnv(
-  "CLOUDFLARE_TURNSTILE_SECRET_KEY",
-  { isRequired: false }
-);
-
-export const DOMAIN = getEnv("DOMAIN", { isRequired: false }); // preview environments need no domain
-export const EXCHANGE_RATES_API_KEY = getEnv("EXCHANGE_RATES_API_KEY", {
-  isRequired: false,
-  isSecret: true
-});
-
-const INNGEST_DEV = getEnv("INNGEST_DEV", { isRequired: false });
-
-export const INNGEST_SIGNING_KEY = getEnv("INNGEST_SIGNING_KEY", {
-  isRequired: !INNGEST_DEV,
-  isSecret: true
-});
-export const INNGEST_EVENT_KEY = getEnv("INNGEST_EVENT_KEY", {
-  isRequired: !INNGEST_DEV,
-  isSecret: true
-});
-
-export const ERP_URL =
-  getEnv("ERP_URL", { isRequired: false, isSecret: false }) ??
-  "https://app.carbon.ms";
-export const MES_URL =
-  getEnv("MES_URL", { isRequired: false, isSecret: false }) ??
-  "https://mes.carbon.ms";
-
-export const GOOGLE_PLACES_API_KEY = getEnv("GOOGLE_PLACES_API_KEY", {
-  isRequired: false
-});
-
-const itarEnvironment = getEnv("CONTROLLED_ENVIRONMENT", {
-  isRequired: false,
-  isSecret: false
-});
-
-export const CONTROLLED_ENVIRONMENT = parseBoolean(itarEnvironment, false);
-
-export const NOVU_APPLICATION_ID = getEnv("NOVU_APPLICATION_ID", {
-  isRequired: false,
-  isSecret: false
-});
-export const NOVU_API_URL =
-  getEnv("NOVU_API_URL", {
-    isRequired: false,
-    isSecret: false
-  }) ?? "https://api.novu.co";
-
-export const NOVU_SECRET_KEY = getEnv("NOVU_SECRET_KEY", {
-  isRequired: false,
-  isSecret: true
-});
-
-export const ONSHAPE_CLIENT_ID = getEnv("ONSHAPE_CLIENT_ID", {
-  isRequired: false
-});
-export const ONSHAPE_CLIENT_SECRET = getEnv("ONSHAPE_CLIENT_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-export const ONSHAPE_OAUTH_REDIRECT_URL = getEnv("ONSHAPE_OAUTH_REDIRECT_URL", {
-  isRequired: false
-});
-
-export const QUICKBOOKS_CLIENT_ID = getEnv("QUICKBOOKS_CLIENT_ID", {
-  isRequired: false
-});
-
-export const QUICKBOOKS_CLIENT_SECRET = getEnv("QUICKBOOKS_CLIENT_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-
-export const QUICKBOOKS_WEBHOOK_SECRET = getEnv("QUICKBOOKS_WEBHOOK_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-
-export const RESEND_DOMAIN =
-  getEnv("RESEND_DOMAIN", {
-    isRequired: false
-  }) ?? "carbon.ms";
-
-export const SLACK_BOT_TOKEN = getEnv("SLACK_BOT_TOKEN", {
-  isRequired: false
-});
-export const SLACK_CLIENT_ID = getEnv("SLACK_CLIENT_ID", {
-  isRequired: false
-});
-export const SLACK_CLIENT_SECRET = getEnv("SLACK_CLIENT_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-export const SLACK_OAUTH_REDIRECT_URL = getEnv("SLACK_OAUTH_REDIRECT_URL", {
-  isRequired: false
-});
-export const SLACK_SIGNING_SECRET = getEnv("SLACK_SIGNING_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-export const SLACK_STATE_SECRET = getEnv("SLACK_STATE_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-
-export const SUPABASE_SERVICE_ROLE_KEY = getEnv("SUPABASE_SERVICE_ROLE_KEY");
-export const SUPABASE_DB_URL = getEnv("SUPABASE_DB_URL", {
-  isSecret: true,
-  isRequired: true
-});
-export const SUPABASE_AUTH_EXTERNAL_AZURE_CLIENT_ID = getEnv(
-  "SUPABASE_AUTH_EXTERNAL_AZURE_CLIENT_ID",
-  {
-    isRequired: false,
-    isSecret: true
-  }
-);
-export const SUPABASE_AUTH_EXTERNAL_GOOGLE_CLIENT_ID = getEnv(
-  "SUPABASE_AUTH_EXTERNAL_GOOGLE_CLIENT_ID",
-  {
-    isRequired: false,
-    isSecret: true
-  }
-);
-
-export const SESSION_SECRET = getEnv("SESSION_SECRET");
-export const SESSION_KEY = "auth";
-export const SESSION_ERROR_KEY = "error";
-export const STRIPE_SECRET_KEY = getEnv("STRIPE_SECRET_KEY", {
-  isRequired: false
-});
-export const STRIPE_WEBHOOK_SECRET = getEnv("STRIPE_WEBHOOK_SECRET", {
-  isRequired: false
-});
-export const STRIPE_BYPASS_COMPANY_IDS = getEnv("STRIPE_BYPASS_COMPANY_IDS", {
-  isRequired: false
-});
-export const STRIPE_BYPASS_USER_IDS = getEnv("STRIPE_BYPASS_USER_IDS", {
-  isRequired: false
-});
-export const GTM_URL = getEnv("GTM_URL", {
-  isRequired: false,
-  isSecret: false
-});
-export const GTM_EVENTS_API_SECRET_KEY = getEnv("GTM_EVENTS_API_SECRET_KEY", {
-  isRequired: false,
-  isSecret: true
-});
-export const REDIS_URL = getEnv("REDIS_URL", {
-  isRequired: true,
-  isSecret: true
-});
-export const SESSION_MAX_AGE = 60 * 60 * 24 * 7; // 7 days;
-export const REFRESH_ACCESS_TOKEN_THRESHOLD = 60 * 10; // 10 minutes left before token expires
-export const VERCEL_URL = getEnv("VERCEL_URL", { isSecret: false });
-
-export const XERO_CLIENT_ID = getEnv("XERO_CLIENT_ID", {
-  isRequired: false
-});
-export const XERO_CLIENT_SECRET = getEnv("XERO_CLIENT_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-export const XERO_WEBHOOK_SECRET = getEnv("XERO_WEBHOOK_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-
-export const JIRA_CLIENT_ID = getEnv("JIRA_CLIENT_ID", {
-  isRequired: false
-});
-export const JIRA_CLIENT_SECRET = getEnv("JIRA_CLIENT_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-export const JIRA_OAUTH_REDIRECT_URL = getEnv("JIRA_OAUTH_REDIRECT_URL", {
-  isRequired: false
-});
-export const JIRA_STATE_SECRET = getEnv("JIRA_STATE_SECRET", {
-  isRequired: false,
-  isSecret: true
-});
-
-/**
- * Shared envs
- */
-
-export const NODE_ENV = getEnv("NODE_ENV", {
-  isSecret: false,
-  isRequired: false
-});
-
-export const VERCEL_ENV =
-  getEnv("VERCEL_ENV", {
-    isSecret: false,
-    isRequired: false
-  }) ?? NODE_ENV;
-
-export const POSTHOG_API_HOST = getEnv("POSTHOG_API_HOST", {
-  isSecret: false
-});
-export const POSTHOG_PROJECT_PUBLIC_KEY = getEnv("POSTHOG_PROJECT_PUBLIC_KEY", {
-  isSecret: false
-});
-export const SUPABASE_URL = getEnv("SUPABASE_URL", { isSecret: false });
-export const SUPABASE_ANON_KEY = getEnv("SUPABASE_ANON_KEY", {
-  isSecret: false
-});
-
-export const DEFAULT_LANGUAGE =
-  getEnv("DEFAULT_LANGUAGE", {
-    isSecret: false,
-    isRequired: false
-  }) ?? "en";
-
-export const RATE_LIMIT = parseInt(
-  getEnv("RATE_LIMIT", { isRequired: false, isSecret: false }) || "5",
-  10
-);
-
-export function getAppUrl() {
-  if (VERCEL_ENV === "production" || NODE_ENV === "production") {
-    return ERP_URL
-      ? ERP_URL
-      : CONTROLLED_ENVIRONMENT
-        ? "https://itar.carbon.ms"
-        : "https://app.carbon.ms";
-  }
-
-  if (VERCEL_ENV === "preview") {
-    return `https://${process.env.VERCEL_URL}`;
-  }
-
-  return "http://localhost:3000";
-}
-
-export function getMESUrl() {
-  if (VERCEL_ENV === "production" || NODE_ENV === "production") {
-    return MES_URL
-      ? MES_URL
-      : CONTROLLED_ENVIRONMENT
-        ? "https://mes.itar.carbon.ms"
-        : "https://mes.carbon.ms";
-  }
-
-  if (VERCEL_ENV === "preview") {
-    return `https://${process.env.VERCEL_URL}`;
-  }
-
-  return "http://localhost:3001";
-}
-
-export function getBrowserEnv() {
-  return {
-    CARBON_EDITION,
-    CARBON_API_URL,
-    CLOUDFLARE_TURNSTILE_SITE_KEY,
-    CONTROLLED_ENVIRONMENT,
-    ERP_URL,
-    GOOGLE_PLACES_API_KEY,
-    JIRA_CLIENT_ID,
-    MES_URL,
-    NODE_ENV,
-    NOVU_APPLICATION_ID,
-    NOVU_API_URL,
-    ONSHAPE_CLIENT_ID,
-    POSTHOG_API_HOST,
-    POSTHOG_PROJECT_PUBLIC_KEY,
-    QUICKBOOKS_CLIENT_ID,
-    SUPABASE_ANON_KEY,
-    SUPABASE_URL,
-    VERCEL_ENV,
-    VERCEL_URL,
-    XERO_CLIENT_ID,
-    DEFAULT_LANGUAGE
-  };
-}
-
-export function isVercel() {
-  return VERCEL_URL?.includes("vercel.app") ?? false;
-}
+export * from "@carbon/env";
