+++ packages/ee/package.json
@@ -32,6 +32,7 @@
     "@carbon/auth": "workspace:*",
     "@carbon/database": "workspace:*",
     "@carbon/kv": "workspace:*",
+    "@carbon/lib": "workspace:*",
     "@carbon/react": "workspace:*",
     "@carbon/utils": "workspace:*",
     "@internationalized/date": "catalog:",
