+++ packages/form/package.json
@@ -31,7 +31,7 @@
     "sst": "catalog:",
     "tiny-invariant": "1.3.3",
     "zod": "catalog:",
-    "zustand": "5.0.12"
+    "zustand": "catalog:"
   },
   "exports": {
     ".": "./src/index.tsx"
