+++ packages/jobs/src/inngest/functions/tasks/model-thumbnail.ts
@@ -1,5 +1,5 @@
-import { SUPABASE_ANON_KEY, SUPABASE_URL, VERCEL_URL } from "@carbon/auth";
 import { getCarbonServiceRole } from "@carbon/auth/client.server";
+import { SUPABASE_ANON_KEY, SUPABASE_URL, VERCEL_URL } from "@carbon/env";
 import { inngest } from "../../client";
 
 export const modelThumbnailFunction = inngest.createFunction(
