+++ packages/jobs/src/inngest/functions/tasks/user-admin.ts
@@ -1,8 +1,8 @@
 import type { Result } from "@carbon/auth";
-import { CarbonEdition, getAppUrl, RESEND_DOMAIN } from "@carbon/auth";
 import { getCarbonServiceRole } from "@carbon/auth/client.server";
 import { deactivateUser } from "@carbon/auth/users.server";
 import { InviteEmail } from "@carbon/documents/email";
+import { CarbonEdition, getAppUrl, RESEND_DOMAIN } from "@carbon/env";
 import { sendEmail } from "@carbon/lib/resend.server";
 import { updateSubscriptionQuantityForCompany } from "@carbon/stripe/stripe.server";
 import { Edition } from "@carbon/utils";
