+++ packages/lib/package.json
@@ -1,11 +1,13 @@
 {
   "name": "@carbon/lib",
   "dependencies": {
-    "@carbon/auth": "workspace:*",
     "@carbon/database": "workspace:*",
+    "@carbon/env": "workspace:*",
     "@carbon/kv": "workspace:*",
+    "@carbon/notifications": "workspace:*",
     "@slack/web-api": "catalog:",
     "@supabase/supabase-js": "catalog:",
+    "inngest": "3.54.0",
     "resend": "3.5.0",
     "sst": "catalog:",
     "supabase": "catalog:",
@@ -22,8 +24,11 @@
     "typescript": "catalog:"
   },
   "exports": {
+    "./events": "./src/events.ts",
+    "./inngest": "./src/inngest/client.ts",
     "./resend.server": "./src/resend.server.ts",
     "./slack.server": "./src/slack.server.ts",
+    "./trigger": "./src/trigger.ts",
     "./twenty.server": "./src/twenty.server.ts"
   },
   "private": true,
