+++ packages/utils/package.json
@@ -2,6 +2,7 @@
   "name": "@carbon/utils",
   "dependencies": {
     "@internationalized/date": "catalog:",
+    "@supabase/supabase-js": "catalog:",
     "@tiptap/react": "catalog:",
     "cookie": "catalog:",
     "dompurify": "catalog:",
