+++ pnpm-workspace.yaml
@@ -1,34 +1,42 @@
 packages:
-  - 'apps/*'
-  - 'ci'
-  - 'examples/*'
-  - 'packages/*'
+  - apps/*
+  - ci
+  - examples/*
+  - packages/*
 
 catalog:
-  # React core
-  react: 18.3.1
-  react-dom: 18.3.1
-  '@types/react': 18.3.28
-  '@types/react-dom': 18.3.7
-
-  # React Router 7
-  react-router: 7.12.0
-  '@react-router/dev': 7.12.0
-  '@react-router/serve': 7.12.0
-  '@react-router/remix-routes-option-adapter': 7.12.0
-  remix-flat-routes: 0.8.5
-
-  # Lingui
-  '@lingui/core': 5.9.4
-  '@lingui/react': 5.9.4
-  '@lingui/macro': 5.9.4
+  '@ai-sdk/openai': 2.0.102
+  '@dnd-kit/core': 6.3.1
+  '@dnd-kit/sortable': 8.0.0
+  '@dnd-kit/utilities': 3.2.2
+  '@emoji-mart/data': 1.2.1
+  '@emoji-mart/react': 1.1.1
+  '@internationalized/date': 3.12.0
   '@lingui/cli': 5.9.4
+  '@lingui/core': 5.9.4
+  '@lingui/macro': 5.9.4
+  '@lingui/react': 5.9.4
   '@lingui/vite-plugin': 5.9.4
-
-  # Tiptap
+  '@nanostores/react': 0.7.3
+  '@react-aria/i18n': 3.12.16
+  '@react-email/components': 0.3.3
+  '@react-pdf/renderer': 3.4.5
+  '@react-router/dev': 7.12.0
+  '@react-router/remix-routes-option-adapter': 7.12.0
+  '@react-router/serve': 7.12.0
+  '@shikijs/langs': 4.0.2
+  '@shikijs/themes': 4.0.2
+  '@shikijs/types': 4.0.2
+  '@slack/web-api': 7.10.0
+  '@supabase/functions-js': 2.1.5
+  '@supabase/postgrest-js': 1.9.2
+  '@supabase/storage-js': 2.5.5
+  '@supabase/supabase-js': 2.80.0
+  '@tailwindcss/container-queries': 0.1.1
+  '@tailwindcss/forms': 0.5.11
+  '@tailwindcss/line-clamp': 0.4.4
+  '@tailwindcss/typography': 0.5.19
   '@tiptap/core': 2.27.2
-  '@tiptap/react': 2.27.2
-  '@tiptap/starter-kit': 2.27.2
   '@tiptap/extension-color': 2.27.2
   '@tiptap/extension-list-item': 2.27.2
   '@tiptap/extension-table': 2.27.2
@@ -39,92 +47,54 @@ catalog:
   '@tiptap/extension-underline': 2.27.2
   '@tiptap/extension-youtube': 2.27.2
   '@tiptap/pm': 2.27.2
-
-  # Shiki
-  '@shikijs/types': 4.0.2
-  '@shikijs/themes': 4.0.2
-  '@shikijs/langs': 4.0.2
-
-  # Supabase
-  '@supabase/supabase-js': 2.80.0
-  '@supabase/functions-js': 2.1.5
-  '@supabase/postgrest-js': 1.9.2
-  '@supabase/storage-js': 2.5.5
-  supabase: 2.89.0
-
-  # Vercel
+  '@tiptap/react': 2.27.2
+  '@tiptap/starter-kit': 2.27.2
+  '@types/react': 18.3.28
+  '@types/react-dom': 18.3.7
   '@vercel/analytics': 1.6.1
   '@vercel/react-router': 1.2.6
-
-  # Tailwind ecosystem
-  tailwindcss: 3.4.6
-  tailwindcss-animate: 1.0.7
-  tailwind-merge: 2.6.1
-  tailwind-scrollbar: 2.1.0
-  tailwind-scrollbar-hide: 1.3.1
-  '@tailwindcss/container-queries': 0.1.1
-  '@tailwindcss/forms': 0.5.11
-  '@tailwindcss/line-clamp': 0.4.4
-  '@tailwindcss/typography': 0.5.19
   autoprefixer: 10.4.14
-  postcss: 8.5.9
-
-  # Vite + tooling
-  vite: 7.3.2
-  vite-plugin-babel-macros: 1.0.6
-  vite-tsconfig-paths: 5.1.4
-  tsup: 8.5.1
-  tsx: 4.21.0
-  typescript: 5.8.3
-  vitest: 4.1.4
-  rimraf: 5.0.10
-  npm-run-all: 4.1.5
-  cross-env: 7.0.3
-  nodemon: 3.1.14
-  wait-on: 7.2.0
-
-  # Validation / forms
-  zod: 3.25.76
-  zod-form-data: 2.0.8
-
-  # Util libs (most-used)
-  nanoid: 5.1.7
+  axios: 1.15.0
+  class-variance-authority: 0.7.1
   cookie: 1.1.1
+  cross-env: 7.0.3
   dompurify: 3.4.0
   framer-motion: 11.18.2
+  intl-parse-accept-language: 1.0.0
   isbot: '5'
+  kysely: 0.28.15
+  nanoid: 5.1.7
+  nanostores: 0.9.5
+  nodemon: 3.1.14
   non.geist: 1.0.4
-
-  # Slack
-  '@slack/web-api': 7.10.0
-
-  # AI SDK
-  '@ai-sdk/openai': 2.0.102
-
-  # Aria / i18n
-  '@react-aria/i18n': 3.12.16
-  '@internationalized/date': 3.12.0
-
-  # SST
-  sst: 4.6.9
-
-  # Misc shared
+  npm-run-all: 4.1.5
+  postcss: 8.5.9
+  posthog-js: 1.367.0
+  react: 18.3.1
+  react-dom: 18.3.1
   react-dropzone: 14.4.1
   react-hotkeys-hook: 4.5.1
   react-icons: 5.6.0
   react-pdf: 10.4.1
   react-resizable-panels: 2.1.9
-  '@react-pdf/renderer': 3.4.5
-  '@dnd-kit/core': 6.3.1
-  '@dnd-kit/sortable': 8.0.0
-  '@dnd-kit/utilities': 3.2.2
-  '@emoji-mart/data': 1.2.1
-  '@emoji-mart/react': 1.1.1
-  '@nanostores/react': 0.7.3
-  nanostores: 0.9.5
-  intl-parse-accept-language: 1.0.0
-  class-variance-authority: 0.7.1
-  kysely: 0.28.15
-  axios: 1.15.0
-  '@react-email/components': 0.3.3
-  posthog-js: 1.367.0
+  react-router: 7.12.0
+  remix-flat-routes: 0.8.5
+  rimraf: 5.0.10
+  sst: 4.6.9
+  supabase: 2.89.0
+  tailwind-merge: 2.6.1
+  tailwind-scrollbar: 2.1.0
+  tailwind-scrollbar-hide: 1.3.1
+  tailwindcss: 3.4.6
+  tailwindcss-animate: 1.0.7
+  tsup: 8.5.1
+  tsx: 4.21.0
+  typescript: 5.8.3
+  vite: 7.3.2
+  vite-plugin-babel-macros: 1.0.6
+  vite-tsconfig-paths: 5.1.4
+  vitest: 4.1.4
+  wait-on: 7.2.0
+  zod: 3.25.76
+  zod-form-data: 2.0.8
+  zustand: 5.0.12
